# lib/storage: update dedup tests